refactor: delegate output tasks to managers
