+++ src/agent/runtime/OutputHandler/types.ts
@@ -1,0 +1,4 @@
+export interface NamedOutputFile {
+  source: string;
+  path: string;
+}
